Multiple Elitisms

+++ src/org/spicydog/FitnessCalc.java
@@ -15,8 +15,8 @@ public class FitnessCalc {
     }
 
     // Calculate inidividuals fittness by comparing it to our candidate solution
-    static int getFitness(Individual individual) {
-        int fitness = 0;
+    static double getFitness(Individual individual) {
+        double fitness = 0;
         // Loop through our individuals genes and compare them to our cadidates
         for (int i = 0; i < individual.size() && i < solution.length; i++) {
             if (individual.getGene(i) == solution[i]) {
@@ -27,7 +27,7 @@ public class FitnessCalc {
     }
 
     // Get optimum fitness
-    static int getMaxFitness() {
+    static double getMaxFitness() {
         int maxFitness = solution.length;
         return maxFitness;
     }
